move definition to cc file from header file
